refactored backend by updating file structure
