+++ backend/controllers/pagesControllers/deletePage.controller.js
@@ -1,8 +1,8 @@
-import Page from '../../models/Page.js';
+import Page from '../../models/Page.model.js';
 import { verifyToken } from '../../util/token.js';
 import { z } from 'zod';
 
-export default async function savePage(req) {
+export default async function deletePage(req) {
   try {
     // Get token from cookies
     const token = req.cookies?.token;
@@ -15,12 +15,11 @@ export default async function savePage(req) {
       };
     }
 
-    // Zod validation for pageId and newPageData only
-    const savePageSchema = z.object({
+    // Zod validation for pageId only
+    const deletePageSchema = z.object({
       pageId: z.string().min(1, 'Page ID is required'),
-      newPageData: z.string().min(0, 'Page data is required'),
     });
-    const parseResult = savePageSchema.safeParse(req.body);
+    const parseResult = deletePageSchema.safeParse(req.body);
     if (!parseResult.success) {
       return {
         resStatus: 400,
@@ -29,9 +28,10 @@ export default async function savePage(req) {
         },
       };
     }
-    const { pageId, newPageData } = parseResult.data;
 
-    // verify user and find page in DB
+    const { pageId } = parseResult.data;
+
+    // verify user is logged in
     const user = await verifyToken(token);
     if (!user) {
       return {
@@ -42,6 +42,7 @@ export default async function savePage(req) {
       };
     }
 
+    // find page in db
     const page = await Page.findById(pageId);
     if (!page) {
       return {
@@ -52,7 +53,7 @@ export default async function savePage(req) {
       };
     }
 
-    // checks if user is the owner of the page
+    // verify user owns the page
     if (!page.owner.equals(user._id)) {
       return {
         resStatus: 403,
@@ -62,15 +63,25 @@ export default async function savePage(req) {
       };
     }
 
-    // change the page data and save to DB
-    page.pageData = newPageData;
-    await page.save();
+    // delete the page
+    const pageDeleted = await Page.findByIdAndDelete(pageId);
+    if (!pageDeleted) {
+      return {
+        resStatus: 500,
+        resMessage: {
+          message: 'Failed to delete page',
+        },
+      };
+    }
 
-    // return new page to user
+    // remove the pageid from the user page array
+    user.pages.pull(page._id);
+    await user.save();
+
     return {
       resStatus: 200,
       resMessage: {
-        'Updated Page': page,
+        message: 'Page deleted',
       },
     };
   } catch (err) {
