+++ backend/routes/pages/pages.routes.js
@@ -1,12 +1,12 @@
 import express from 'express';
 
-import createPage from './createPage.js';
-import savePage from './savePage.js';
-import renamePage from './renamePage.js';
-import { getPage, getPages } from './getPages.js';
-import deletePage from './deletePage.js';
-import sharePage from './sharePage.js';
-import { publicShare, getPublicShare } from './publicShare.js';
+import createPage from '../../controllers/pagesControllers/createPage.controller.js';
+import savePage from '../../controllers/pagesControllers/savePage.controller.js';
+import renamePage from '../../controllers/pagesControllers/renamePage.controller.js';
+import { getPage, getPages } from '../../controllers/pagesControllers/getPages.controller.js';
+import deletePage from '../../controllers/pagesControllers/deletePage.controller.js';
+import sharePage from '../../controllers/pagesControllers/sharePage.controller.js';
+import { publicShare, getPublicShare } from '../../controllers/pagesControllers/publicShare.controller.js';
 
 const router = express.Router();
